Instruction to configure swing available
use 'swing beatRatio [beatRatio ...] ["..." phraseLength]' to apply
swing to a phrase (a series of beats). If no phraseLength argument
is provided, it is implied to be beatRatioList.length + 1.
The beatRatio list provides the ratios to adjust the duration of each
beat in a phrase. Beats at the end of the phrase without ratios
specified are implied to share remaining time evenly.

For example:
```
swing 1.5
4/4
```
will play 4 of the following lengths:
1.5, .5, 1.5, .5

If we specify a longer phrase:
```
swing 1.75 ... 4
4/4
```
will play 4 beats of the following lengths:
1.75, .75, .75, .75

Swing ratios smaller than 1 are possible as well.

Additionally, this introduced a stricter handling of numbers in the parser:
we distinguish whole numbers and certain values (the top number in a measure,
or the length of a phrase) require whole number arguments.