Support for additive meters
For example, instead of '7/4', one can write '4+3/4' to denote that we
emphasize the 4th beat.

This required a breaking change: the syntax for declaring bpm is less
natural, because we now put 'bpm' before the number. I might later look
for a way to specify the parsing such that putting 'bpm' after the
number reduces conflicts, but for now this is good enough.