polyrhythm works.
Denote like so:
4:3/4
to play 4/4 with three beats played evenly underneath
or 4:5/3 to play 5 beats underneath.
Or do both:
4:5:3/4 is valid.
As is
4:2+3:3/4
and so forth.

It even doesn't break with swing, though it also
doesn't make tons of sense with swing. There's not really a good way
to notate swung polyrhythms, or perform them or think about them
for that matter.