Name and reuse sections
Introduces an assigment operator '=' and a special 'RE' instruction.
Assignment has the syntax `IDENTIFIER = [section]`, for example `test =
4/4` saves a single bar of 4/4 to the identifier 'test'. Later, the
identifier can be used alone in place of a section, e.g.
```
test = 4/4
test * 2
```
to play 2 bars of 4/4.
The parameters (presently just BPM and SWING) are saved alongside the
section and are used to play it back, so
```
bpm 120
test = 4/4
bpm 240
test * 2
```
still plays those 2 bars of 4 at 120bpm.
However, the 'RE' keyword can be used to reinterpret the section with
the current parameters. (And fails if that is not possible - note the
swing param is finnicky).

```
bpm 120
test = 4/4
bpm 200
test
re test * 3
```
plays one bar of 4 at 120 bpm, then 3 bars at 200bpm.

If the section defines changes in bpm, etc, those are still respected
when it is reinterpreted. For example:

```
bpm 120
test = (
  4/4
  bpm 200
  4/4
)

bpm 160
test
re test
```
plays 1 bar of 4 at 120bpm, the next at 200bpm, the next at 160bpm, and
the last at 200bpm.