Audio separated from sections/timing information
Rather than rely on the sections (ergo beaterator, etc) to provide the
audio players and recorders use, sections supply only scheduling and
emphasis information to players and recorders. These orchestrate audio
by passing the emphasis information to an ensemble, which in turn
provides the buffer to play. The downstream benefit here is that
ensembles can be swapped out, allowing easy customization of click
sounds later on.