Accelerando and ritardando support added
Using either the 'ACC' or 'RIT' keyword - and no, it doesn't actually
matter which keyword is used, regardless the particulars of the tempo
change - users can schedule gradual BPM changes, optionally with a
delay.

```
bpm 120
acc 240 16
4/4 * 8
```
sets the track to accelerate to 240 bpm from the current bpm, reaching
that pace in 16 beats, and continuing to play at 240 bpm afterward for
16 more beats.

Relative bpms are supported, so
```
bpm 120
acc * 2 16
4/4 * 8
```
is equivalent to the above.

The start of the change can be delayed:
```
bpm 120
acc 240 16 4
4/4 * 8
```
will delay the start of the temp change four beats. However, it will
still end by the 16th beat after the accelerando is defined, i.e. the
above plays 4 beats at 120bpm, accelerates to 240 bpm over 12 beats, and
plays 16 beats at 240bpm.

'RIT' is a synonym for 'ACC', though stylistically you should use 'RIT'
when slowing down and 'ACC' when speeding up.