parser always reaches end of file before returning
It was possible to create tracks that 'end' in the middle and had free
text afterward by placing certain _valid_ symbols (e.g., ')') after a
valid section. The section would be returned and the track would end,
but the text could continue. Such a small typo is easy to miss, and
could leave users wondering why a track appeared to stop in the middle.

The parser can no longer return unless an end of file is reached.

List handling simplified as well -- all push, no slice.