`set`, `aTempo` on BPM  after acc/rit no longer corrupts bpm stack
When a tempo adjustment is complete, the next call on a BPM object that
sets or retrieves the bpm is supposed to clear the tempo adjustment and
push the target bpm onto the BPM stack, as that is the current BPM at
that point - _then_ the call map push, pop, get return the value at the
top of the bpm stack. However, this only worked correctly for 'get'.
Other calls used `assertNotAdjusting`, which erroneously pushed
'undefined' onto the stack instead of the target BPM. That corrupted the
stack and subsequent get calls could fail.