AST reports exact text locations.
Location and Range classes added to describe places and ranges of source
code. The parser binds raw range data into the AST describing the exact
line and column of starts and ends for every symbol and group thereof
(i.e., every AST node). The interpreter uses this data to attach ranges
to the various sections defined. It futzes with this sometimes, such
that (for example), a repeating section like `4/4 * 3` is defined with
the _first_ section produced having a range describing _just_ the first
three characters - '4/4' - but all subsequent sections described with
the full 6 character range, including ' * 3', representing the fact that
the repeatings are _the whole thing_, but that the initial pass is
interesting on its own.

This will support many advanced features, such as code suggestions,
inline error messages, playback of ranges, and marks over the text
showing current playback position.