Player can respond to selections
The interpreter produces more complex track data:
1. sections within assignments (e.g. the `4/4` in `foo = 4/4`) are
also returned alongside the full track, with the locations thereof
reported.
2. When a track is to be played, a selection is optionally provided. The
track selects a section to be played by looking for a section (be it the
whole track, or a subsection in an assignment) that wholly contains the
selection, falling back to the whole track. (This allows text in an
assignment to be selected and played back even though when the whole
track is played back, the definition is not played.)
3. On playback, the beaterator skips beats that are not defined within
the selection. Ergo, a range of input program can be given to the player
to decide what, if anything, is played back.