Ranges tightly coupled to tracks
Before, the track produced by the interpreter was a handful of sections,
and it could later be combined with a range to produce a playable
section + range. Now, the interpreter is passed the selection at parse
time, and produces a track with a single section and a range. In this
way, the track is a self-contained program produced by the interpreter
after a single call from the editor. This helps decouple the etcnome
language from the editor, which now knows even less about what it's
actually editing.